+++ client/src/error.rs
@@ -206,6 +206,19 @@ pub enum Error {
     UrlValidationError(String),
 
     //////////////////////////////////////////////////////////////////////
+    // Participation
+    //////////////////////////////////////////////////////////////////////
+    /// Invalid participations error
+    #[cfg(feature = "participation")]
+    #[error("invalid participations")]
+    InvalidParticipations,
+    /// IO error
+    #[cfg(feature = "participation")]
+    #[error("`{0}`")]
+    #[serde(serialize_with = "display_string")]
+    IoError(#[from] std::io::Error),
+
+    //////////////////////////////////////////////////////////////////////
     // Ledger Nano
     //////////////////////////////////////////////////////////////////////
     /// Denied by User