+++ client/src/node_api/participation/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! Node participation API.
+//! https://github.com/iota-community/treasury/blob/main/specifications/hornet-participation-plugin.md
+
+pub mod responses;
+pub mod routes;
+pub mod types;