+++ client/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `OutputMetadata::new()` and all field getters;
+- `participation` feature with routes, responses and types;
 
 ### Changed